requirements.txt
beautifulsoup4==4.8.1
boto3==1.17.112
botocore==1.20.112
Django==2.2.28
django-bootstrap4==1.0.1
django-ckeditor==5.7.1
django-jquery==3.1.0
django-jquery-lightbox==2.6
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-storages==1.8
jmespath==0.10.0
Pillow==6.2.1
python-dateutil==2.8.2
pytz
s3transfer==0.4.2
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.26.6
